backend/requirements.txt
Django==4.0.4
django-filter==2.4.0
djangorestframework==3.12.4
django-cors-headers==3.10.0
daphne==3.0.2
pytest-django==4.5.2
Pillow==9.5.0
django-modeltranslation>=0.17.3
beautifulsoup4==4.10.0
cssutils==2.3.0
pytz==2021.3
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
model-bakery==1.9.0
drf-spectacular==0.25.1
django-ordered-model==3.7.4
docker==6.1.2
django-ordered-model==3.7.4
urllib3<2
git+https://${GITHUB_API_KEY}@github.com/Smart-Study-Graz/django-utils@0.7.0#egg=django_utils
git+https://${GITHUB_API_KEY}@github.com/Smart-Study-Graz/smartstudy_auth@3.0.1#egg=smartstudy_auth
git+https://${GITHUB_API_KEY}@github.com/Smart-Study-Graz/django-smartstudy-org-core@0.6.0#egg=smartstudy_org_core
git+https://${GITHUB_API_KEY}@github.com/Smart-Study-Graz/django-smartstudy-assets@0.12.0#egg=smartstudy_assets
git+https://${GITHUB_API_KEY}@github.com/Smart-Study-Graz/py-scorm@0.2.3#egg=py-scorm
